+++ agents/ghost.js
@@ -3,7 +3,7 @@ import { soulseek } from './souls.js';
 
 import { getTimeOfDayEmoji } from '../tools/time-of-day-emoji.js';
 
-import AIServiceManager from '../ai-services.js';
+import AIServiceManager from '../tools/ai-services.js';
 const ai = new AIServiceManager();
 await ai.useService('groq');
 await ai.updateConfig({
