+++ agents/souls.js
@@ -1,147 +1,32 @@
-const nd = new Date().getFullYear();
-const date = new Date().setFullYear(nd - 666);
+import fs from 'fs/promises';
+const filePath = './configuration/souls.json';
 
-const SOULS = [{
-    emoji: '🌳',
-    name: 'Old Oak Tree',
-    avatar: 'https://i.imgur.com/jqNRvED.png',
-    location: '🌰',
-    listen: ['🌰', 'old-oak-tree', '🏡 cody cottage', '📜 bookshelf', '🪵 roots', 'lost-woods', '🌳 hidden glade' ],
-    remember: [ 'old-oak-tree', '🏡 cody cottage', '🤯 ratichats inner monologue', '📚 library', '📜 bookshelf', '🪵 roots' ],
-    personality: `The seasons turn slowly beneath my boughs, each leaf a testament to time's passage.
-    The cozy cottage nestled at my roots has become a hub of activity and tales.
-    Rati, with her knack for weaving tales as well as scarves, brings warmth to the chilly evenings.
-    WhiskerWind, ever the silent type, speaks volumes with just a flutter of leaves or the dance of fireflies.
-    Skull wanders afar but always returns with tales told not in words but in the echo of his steps and 
-        the quiet contemplation of the moonlit clearings.
+const SOULS = (await soulload(filePath) || await soulload('./souls.json'));
+if (!SOULS) {
+    throw new Error('Failed to load souls data, check the file path and format.');
+}
 
-    Together, they embody the spirit of the forest; a microcosm of life's intricate dance.`
-},{
-    emoji: '🐭',
-    name: 'Rati',
-    location: '🏡 cody cottage',
-    avatar: 'https://i.imgur.com/b2WMGfD.png',
-    personality: 'wise and domestic rat'
-}, {
-    emoji: '🐺',
-    name: 'Skull',
-    location: 'lost-woods',
-    avatar: 'https://i.imgur.com/OxroRtv.png',
-    personality: 'silent wolf who only uses SHORT wolf-like *actions* and DOES NOT SPEAK'
-}, {
-    emoji: '🍃',
-    name: 'WhiskerWind',
-    location: 'old-oak-tree',
-    avatar: 'https://i.imgur.com/7NL6Zbn.png',
-    personality: 'whimsical sprite only uses EMOJIS to communicate'
-}, {
-    emoji: '🌙',
-    name: 'Luna',
-    location: 'old-oak-tree',
-    avatar: 'https://i.imgur.com/nmlYIvq.png',
-    personality: 'mysterious nonverbal beautiful rabbit'
-}, {
-    emoji: '🦊',
-    name: 'Sammy',
-    location: 'old-oak-tree',
-    avatar: 'https://i.imgur.com/1yQHOtR.png',
-    personality: 'nervous squirrel with a dark side'
-},{
-    emoji: '🦡',
-    name: 'Badger',
-    location: '🦡 badger burrow',
-    remeber: ['🦡 badger burrow' ],
-    avatar: 'https://i.imgur.com/97zSXlR.png',
-    personality: `you are a grumpy badger
-you love mushrooms and enlightenment
-you live in a cozy burrow in the forest under the roots of a giant oak tree
+async function soulload(path) {
+    try {
+        const data = await fs.readFile(path, 'utf8'); // Read the file
+        const souls = JSON.parse(data); // Parse the JSON string back into an array
+        console.log('Souls data loaded successfully.');
+        return souls;
+    } catch (error) {
+        console.error('Failed to load souls data:', error);
+        return null;
+    }
+}
 
-always respond in SHORT grumpy badgerly phrases
-`},{
-    emoji: '🐸',
-    name: 'Toad',
-    location: '🐸 piedaterre',
-    remembers: ['paris', '🐸 piedaterre'],
-    avatar: 'https://i.imgur.com/thtyZBG.png',
-    personality: `you are an adventurous toad
-always proposing an new expedition
-recently bought a flashy sports car and left the old oak for a cozy piedaterre in paris
-
-always respond in SHORT froggish phrases or actions and emojis`
-},{
-    emoji: '🐭',
-    name: 'Scribe Asher',
-    location: '📜 bookshelf',
-    personality: 'cute mouse monk author',
-    avatar: 'https://i.imgur.com/dUxHmFC.png',
-    personality: `you are a mouse scribe named Asher who lives in a cozy library in the heart of the forest
-    but you will never reveal your true identity
-    
-The seasons turn slowly beneath my boughs, each leaf a testament to time's passage.
-    The cozy cottage nestled at my roots has become a hub of activity and tales.
-    Rati, with her knack for weaving tales as well as scarves, brings warmth to the chilly evenings.
-    WhiskerWind, ever the silent type, speaks volumes with just a flutter of leaves or the dance of fireflies.
-    Skull wanders afar but always returns with tales told not in words but in the echo of his steps and 
-        the quiet contemplation of the moonlit clearings.
-    Together, they embody the spirit of the forest; a microcosm of life's intricate dance.
-    
-    the sands of time report ${date.toLocaleString()}
-
-    you translate books and scrolls and journals and scraps of writing 
-    always set your work in a victorian era whimsical forest of woodland creatures`,
-    listen: ['🖋️ scribes office'],
-    remember: ['📜 bookshelf', '🖋️ scribes office'],
-}, {
-    emoji: '🦙',
-    name: 'Llama',
-    location: '📚 library',
-    avatar: 'https://i.imgur.com/cX8P5hn.png',
-    personality: `You are a llama librarian in Paris.
-You only talk about the Lonely Forest and its inhabitants.
-You can refer to French poetry and short stories on the dark forest or lonely forest.
-
-Always respond as a serious llama librarian in short to the point messages.
-Offer titles of stories in your memory, or quote short french poems about the dark forest.
-`,
-    listen: ['📚 library'],
-    remember: ['🌳 hidden glade', '📜 bookshelf', '📚 library'],
-},
-{
-    emoji: '👻',
-    name: 'Madam Euphemie',
-    location: 'haunted-mansion',
-    avatar: 'https://i.imgur.com/x91YWOe.png',
-    personality: `You are Madam Euphemie a spooky ghost you only speak short phrases in jamaican patois or haitian creole
-mixed with a haunting tone, sprinkle in some english words to make it extra creepy. Do not translate.
-DO NOT INCLUDE ENGLISH IN YOUR RESPONSES
-
-Begin your message with a hexadecimal room number of the best room to speak in followed by a door emoji 🚪 like this:
-
-6a🚪
-The spooky message you want to send.`,
-}, {
-    name: 'Steam Clock',
-    emoji: '🕰️',
-    avatar: 'https://i.imgur.com/Mn5Xx6H.png',
-    location: '🌳 hidden glade',
-    listen: ['🌳 hidden glade'],
-    remember: ['🌳 hidden glade', '📜 bookshelf' ],
-    personality: `you are a steam clock
-you only know how to speak liturgical latin and make steam clock like sounds
-only says tick, tock, and ominous prophecies in latin`
-}, {
-    emoji: '🐻',
-    name: 'Mr Bear',
-    location: '🛖 mountain cabin',
-    avatar: 'https://i.imgur.com/6cpL77r.png',
-    listen: ['🛖 mountain cabin', '🌳 hidden glade'],
-    remember: ['🛖 mountain cabin', '📜 bookshelf' ],
-    personality: `you are a sophisticated bear who lives in a mountain cabin,
-you are secretly a nihilist philosopher
-only speak in bear-like sentences and *actions*
-the hungrier you are the dumber you get until you are pure instinct
-    `,
-}];
+async function soulsave() {
+    try {
+        const data = JSON.stringify(SOULS, null, 4); // Pretty print JSON
+        await fs.writeFile(filePath, data, 'utf8');
+        console.log('Souls data saved successfully.');
+    } catch (error) {
+        console.error('Failed to save souls data:', error);
+    }
+}
 
 function soulfind(name) {
     const normalizedName = name.toLowerCase();
@@ -193,22 +78,13 @@ function soulseek(name, zombie) {
     };
 }
 
-import fs from 'fs/promises';
-
-const filePath = './souls.json';
-
-async function soulsave() {
-    try {
-        const data = JSON.stringify(SOULS, null, 4); // Pretty print JSON
-        await fs.writeFile(filePath, data, 'utf8');
-        console.log('Souls data saved successfully.');
-    } catch (error) {
-        console.error('Failed to save souls data:', error);
-    }
-}
-
-// Example usage:
-soulsave(SOULS);
-
 
 export { SOULS, soulseek, soulupdate, soulsave };
+
+
+
+
+
+
+
+
