+++ ai-services/replicate.js
@@ -1,7 +1,7 @@
 import Replicate from "replicate";
 
-import AIService from "../ai-service.js";
-import c from "../configuration.js";
+import AIService from "../tools/ai-service.js";
+import c from "../tools/configuration.js";
 const configuration = c('replicate', {
     token: "YOUR_API_KEY",
     model: "meta/meta-llama-3-70b-instruct",
