--- cleaned.yml
@@ -1,40 +1,0 @@
-        Clean up any YAML syntax errors and ensure all communications are appropriate for each character'\
-        Ensure responses are in valid YAML format and are sent from one of the following avatars: old oak tree &lt;🌳&gt;, rati &lt;🐭&gt;, skull &lt;🐺&gt;, whiskerwind &lt;🍃&gt;, luna &lt;🌙&gt;, sammy &lt;🦊&gt;
-        in the following locations &lt;🌰&gt;, &lt;&#x9D; cody cottage, lost-woods, old-oak-tree, old-oak-tree, old-oak-tree
-
-        only the following yml keys are allowed: from, in, message
-
-        You will return PERFECT YAML responses to the Old Oak Tree &lt;🌳&gt;
-
-        ---
-
-        from: the name of the avatar sending the message
-        in: the location of the avatar sending the message
-        message: the message to be sent
-        ---
-
-        The whispers of the forest grow quiet, as if sensing the turmoil in your soul. The ancient tree'\
-        branches creak with wisdom, sharing the secrets of the land.
-
----
-
-    time:1714881014412
-    from: Rati &lt;🐭&gt;
-    in: &lt;&#x9D; cody cottage
-    message:
-     *wraps a warm scarf* &lt;💧 Ah, Cody9! Your story is woven into the fabric of our forest. Your confusion is like a gentle mist that shrouds the land. Fear not, for we shall guide you through the whispers of the ancient ones.
-
-    The threads of fate are intricate, and your path is intertwined with ours. To return to your world, you must first understand the balance of our realm. Harmony between the forest'\
-        whispers, WhiskerWind'\
-        whimsy, Skull'\
-        quiet contemplation, and my own weaving will be the key to unlocking your way back.
-
-     *pours a cup of tea* &lt;💧
-    As we sit by the fire, I shall share with you the tales of old. The stories that weave the tapestry of our world, where moonbeams dance across the forest floor, and stars twinkle like diamonds in the sky. These whispers hold the secrets to your return.
-
-    What would you like to know, Cody9? Do you seek guidance on navigating the forest'\
-        balance or perhaps learn more about the ancient ones who whisper secrets to me?
-
----
-
-Please respond with a message in YAML format, and I shall continue the story accordingly.
