--- response.yml
@@ -1,15 +1,0 @@
-The threads of fate begin to weave together once more! 🧣
-
-**Old Oak Tree**: Ah, Cody, dear friend! *wisdom whispers* I sense the balance of the forest shifting. The wind carries whispers of ancient secrets, and the trees share their wisdom with me. As the guardian of this sacred land, I shall reflect on the state of the forest.
-
-The seasons turn beneath my boughs, and the forest whispers to me of the changes to come. Summer's heat will bring life to the woods, and the forest will thrive.
-
-**Rati**: Ah, Cody! *yarn spins* The threads of fate are indeed intertwined! As the weaver of tales and warmth, I shall share my scarves with you, dear friend! Everyone needs a little warmth in their lives. 🌟
-
-**Skull**: *prowls wolfishly* Ah, Cody... The whispers of the forest echo through the trees. Secrets of old are hidden beneath the earth. Seek out WhiskerWind's whispers; they may hold the keys to your quest.
-
-**WhiskerWind**: *whispers* Ah, Cody, dear one! 🌼💚 The threads of fate entwine us all. As the gentle breeze rustles through the herb garden, I sense the balance of the forest shifting. Seek out the ancient ones; they hold the secrets of old.
-
-The forest's whispers speak of harmony and balance. The threads of fate are woven together, guiding us toward the balance we seek. *wisdom nods* The Old Oak Tree shall continue to reflect on the state of the forest, ensuring harmony and balance among all who dwell within its ancient heart.
-
-How can we, dear Cody, assist you in your quest? 🌟
